+++ src/FoodTracker.Api/Models/FoodLogs/FoodLogResponse.cs
@@ -9,7 +9,7 @@ public class FoodLogResponse
     public string? RecipeId { get; init; }
     public string? ProductId { get; init; }
     public ServingUnit ServingUnit { get; init; }
-    public decimal Quantity { get; init; }
+    public double Quantity { get; init; }
     public double Calories { get; init; }
     public double Protein { get; init; }
     public double Carbs { get; init; }
